Refactoring of MainDialog, flightBookingRecognizer and tests
