+++ libraries/testbot/dialogs/mainDialog.js
@@ -3,6 +3,7 @@
 
 const { TimexProperty } = require('@microsoft/recognizers-text-data-types-timex-expression');
 const { InputHints } = require('botbuilder');
+const { LuisRecognizer } = require('botbuilder-ai');
 const { ComponentDialog, DialogSet, DialogTurnStatus, TextPrompt, WaterfallDialog } = require('botbuilder-dialogs');
 
 const MAIN_WATERFALL_DIALOG = 'mainWaterfallDialog';
@@ -74,11 +75,69 @@ class MainDialog extends ComponentDialog {
     async actStep(stepContext) {
         let bookingDetails = {};
 
-        if (this.luisRecognizer) {
+        if (this.luisRecognizer.isConfigured()) {
             // Call LUIS and gather any potential booking details.
             // This will attempt to extract the origin, destination and travel date from the user's message
             // and will then pass those values into the booking dialog
-            bookingDetails = await this.luisRecognizer.executeLuisQuery(this.logger, stepContext.context);
+            const recognizerResult = await this.luisRecognizer.executeLuisQuery(this.logger, stepContext.context);
+            
+            const getBookingDetails = (recognizerResult) => {
+                const bookingDetails = {};
+                const intent = LuisRecognizer.topIntent(recognizerResult);
+    
+                bookingDetails.intent = intent;
+                bookingDetails.unsupportedCities = [];
+    
+                if (intent === 'Book_flight') {
+                    // We need to get the result from the LUIS JSON which at every level returns an array
+                    bookingDetails.destination = parseCompositeEntity(recognizerResult, 'To', 'Airport');
+                    bookingDetails.origin = parseCompositeEntity(recognizerResult, 'From', 'Airport');
+    
+                    if(!bookingDetails.destination) {
+                        bookingDetails.unsupportedCities.push(getUnsupportedCity(recognizerResult, 'To'))
+                    }
+    
+                    if(!bookingDetails.origin) {
+                        bookingDetails.unsupportedCities.push(getUnsupportedCity(recognizerResult, 'From'))
+                    }
+    
+                    // This value will be a TIMEX. And we are only interested in a Date so grab the first result and drop the Time part.
+                    // TIMEX is a format that represents DateTime expressions that include some ambiguity. e.g. missing a Year.
+                    bookingDetails.travelDate = parseDatetimeEntity(recognizerResult);
+                }
+                return bookingDetails;
+            }
+
+            const parseCompositeEntity = (result, compositeName, entityName) => {
+                const compositeEntity = result.entities[compositeName];
+                if (!compositeEntity || !compositeEntity[0]) return undefined;
+        
+                const entity = compositeEntity[0][entityName];
+                if (!entity || !entity[0]) return undefined;
+        
+                const entityValue = entity[0][0];
+                return entityValue;
+            }
+        
+            const parseDatetimeEntity = (result) => {
+                const datetimeEntity = result.entities['datetime'];
+                if (!datetimeEntity || !datetimeEntity[0]) return undefined;
+        
+                const timex = datetimeEntity[0]['timex'];
+                if (!timex || !timex[0]) return undefined;
+        
+                const datetime = timex[0].split('T')[0];
+                return datetime;
+            }
+        
+            const getUnsupportedCity = (result, compositeName) => {
+                const cityEntity = result.luisResult.entities.find(entity => {
+                  return entity.type = compositeName;
+                });
+                return cityEntity.entity.toUpperCase();
+            }
+
+            bookingDetails = getBookingDetails(recognizerResult);
 
             if(bookingDetails.unsupportedCities && bookingDetails.unsupportedCities.length) {
                 const messageText = `Sorry but the following airports are not supported: ${bookingDetails.unsupportedCities.join(', ')}`;
